+++ tests/system/summer_winter.py
@@ -160,12 +160,12 @@ class TestSummerWinter(tests.helper.test_case_base.TestCaseBase):
 			tests.helper.oh_item.assert_value("Unittest_Summer", "ON")
 			self.assertEqual(1, last_check_mock.oh_send_command.call_count)
 
-			# already summer (no update should be sent)
+			# already summer
 			is_summer_mock.return_value = True
 			with unittest.mock.patch.object(self._summer_winter, "_item_summer") as summer_item:
 				summer_item.value = "ON"
 				self._summer_winter._cb_update_summer()
-				summer_item.oh_send_command.assert_not_called()
+				summer_item.oh_send_command.assert_called_once()
 				self.assertEqual(2, last_check_mock.oh_send_command.call_count)
 
 			# switch back to winter
@@ -174,21 +174,21 @@ class TestSummerWinter(tests.helper.test_case_base.TestCaseBase):
 			tests.helper.oh_item.assert_value("Unittest_Summer", "OFF")
 			self.assertEqual(3, last_check_mock.oh_send_command.call_count)
 
-			# already winter (no update should be sent)
+			# already winter
 			is_summer_mock.return_value = False
 			with unittest.mock.patch.object(self._summer_winter, "_item_summer") as summer_item:
 				summer_item.value = "OFF"
 				self._summer_winter._cb_update_summer()
-				summer_item.oh_send_command.assert_not_called()
+				summer_item.oh_send_command.assert_called_once()
 				self.assertEqual(4, last_check_mock.oh_send_command.call_count)
 
-			# already winter (no update should be sent) | no last_check item -> send_command should not be triggered
+			# already winter | no last_check item -> send_command should not be triggered
 			is_summer_mock.return_value = False
 			with unittest.mock.patch.object(self._summer_winter, "_item_summer") as summer_item:
 				last_check_mock.__bool__.return_value = False
 				summer_item.value = "OFF"
 				self._summer_winter._cb_update_summer()
-				summer_item.oh_send_command.assert_not_called()
+				summer_item.oh_send_command.assert_called_once()
 				self.assertEqual(4, last_check_mock.oh_send_command.call_count)
 
 		# exception from __is_summer